backend/pyproject.toml
[project]
name = "eth_crawler"
version = "0.1.0"
description = "Tracelabs tech assestment"
authors = [
    { name = "Juan Villarejo", email = "jbtwist92@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.119.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.14.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
    "web3>=7.13.0",
]

[tool.ruff]
line-length = 79
target-version = "py314"
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "B",    # flake8-bugbear
    "I",    # isort
    "UP",   # pyupgrade
    "C90",  # mccabe complexity
    "T201",   # print statements are not allowed
]
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    ".git",
]
fix = true
